:zap: Beautify and robustify (fixes #1 :tada:)
Beautify because interface changes and new goofy filenames.
(Having a separate file for css is cool I guess.)
Oh also we robustified the fitness function.
The fitness function returns 0 if either color in the
palette has a sat or lightness that's "out of range".
(This demotes colors that look too light or too dark.)
Oh, also we fixed our first issue yayyy :tada: